requirements.txt
line-bot-sdk
flask
gunicorn
requests
json5